backend_service/requirements.txt
# Backend Service Dependencies

# FastAPI and server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Database
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
neo4j==5.14.1

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP Client for AI Service
httpx==0.25.2

# Rate Limiting
slowapi==0.1.9
redis==5.0.1

# Environment variables
python-dotenv==1.0.0
pyyaml==6.0.1

# Utilities
pydantic==2.5.0
pydantic[email]==2.5.0

